renames file upload util; consolidates callbacks into one place
